create PersonalInfoForm and move all textfields to there
